Final: Semantic slug, refactored components, and CI stability
- Renamed post to 2026-04-19-wcs-travel-pack.md for better semantics.
- Extracted CompactAffiliateLink component for better maintainability.
- Fixed Markdown image pathing with ASSET_PREFIX in MarkdownRenderer.
- Resolved Playwright strict mode violations.
- Verified with audits, linting, and E2E tests.
