+++ src/features/journal/components/BlogPostDetail.tsx
@@ -1,10 +1,10 @@
-import { Share2, ExternalLink, ArrowRight } from 'lucide-react';
-import { Link } from 'react-router-dom';
+import { Share2 } from 'lucide-react';
 import { Box, Stack, Text, Grid } from '@/layouts/Primitives';
 
 import { DetailLayout } from '@/components/layout/DetailLayout';
 import { AffiliateDisclosure } from '@/components/ui/AffiliateDisclosure';
 import { AffiliateCard } from '@/components/ui/AffiliateCard';
+import { CompactAffiliateLink } from '@/components/ui/CompactAffiliateLink';
 import { Post } from '@/lib/content';
 import { affiliateManager } from '@/lib/affiliateManager';
 
@@ -75,42 +75,9 @@ export function BlogPostDetail({ post, onBack, backLabel }: BlogPostDetailProps)
             {/* Remaining items (Compact list) */}
             {remainingAffiliates.length > 0 && (
               <Stack gap={3} marginTop={2}>
-                {remainingAffiliates.map(link => {
-                  const href = affiliateManager.resolveResourceHref({ id: link.id });
-                  const isExternal = href.startsWith('http');
-
-                  return (
-                    <Box
-                      key={link.id}
-                      as={isExternal ? 'a' : Link}
-                      {...(isExternal ? { href, target: '_blank', rel: 'noopener noreferrer sponsored' } : { to: href })}
-                      display="flex"
-                      align="center"
-                      justify="between"
-                      paddingX={4}
-                      paddingY={3}
-                      radius="md"
-                      border
-                      className="group/item hover:border-accent transition-colors bg-surface-alt/10"
-                    >
-                      <Stack direction="row" align="center" gap={3}>
-                        <Box padding={1.5} radius="sm" surface="muted" className="group-hover/item:text-accent transition-colors">
-                          <Text variant="mono" size="xs" weight="font-bold">
-                            {link.category?.toUpperCase() || 'GEAR'}
-                          </Text>
-                        </Box>
-                        <Text variant="body" size="sm" weight="font-bold" className="group-hover/item:text-accent transition-colors">
-                          {link.name}
-                        </Text>
-                      </Stack>
-                      {isExternal ? (
-                        <ExternalLink className="w-3.5 h-3.5 text-dim group-hover/item:text-accent transition-colors" />
-                      ) : (
-                        <ArrowRight className="w-3.5 h-3.5 text-dim group-hover/item:text-accent transition-colors" />
-                      )}
-                    </Box>
-                  );
-                })}
+                {remainingAffiliates.map(link => (
+                  <CompactAffiliateLink key={link.id} link={link} />
+                ))}
               </Stack>
             )}
           </Stack>
